refactor(vet): 프로필 관리 페이지에 따른 edit 폴더 생성 후 폴더구조 변경
